# improve copy, parsing and other performance