--- a/include/adm/elements/private/auto_parent.hpp
+++ b/include/adm/elements/private/auto_parent.hpp
@@ -1,13 +1,23 @@
 #pragma once
 
+/// check and update the parents of self and other
+///
+/// if only one has no parent, set it to the parent of the other
+///
+/// returns true if they both have the same parent after this (or both have no
+/// parent)
 template <typename T1, typename T2>
-void autoParent(T1 t1, T2 t2) {
-  auto parent1 = t1->getParent().lock();
-  auto parent2 = t2->getParent().lock();
-  if (parent1 && !parent2) {
-    parent1->add(t2);
+bool autoParent(T1& self, const std::shared_ptr<T2>& other) {
+  auto selfParent = self.getParent().lock();
+  auto otherParent = other->getParent().lock();
+  if (selfParent && !otherParent) {
+    selfParent->add(other);
+    return true;
   }
-  if (!parent1 && parent2) {
-    parent2->add(t1);
+  if (!selfParent && otherParent) {
+    otherParent->add(self.shared_from_this());
+    return true;
   }
+
+  return selfParent == otherParent;
 }