… include/adm/detail/hex_values.hpp
… include/adm/private/copy.hpp
… src/CMakeLists.txt
… src/detail/hex_values.cpp
… src/document.cpp
… src/elements/audio_block_format_id.cpp
… src/elements/audio_channel_format.cpp
… src/elements/audio_channel_format_id.cpp
… src/elements/audio_content.cpp
… src/elements/audio_content_id.cpp
… src/elements/audio_object.cpp
… src/elements/audio_object_id.cpp
… src/elements/audio_pack_format.cpp
… src/elements/audio_pack_format_id.cpp
… src/elements/audio_programme.cpp
… src/elements/audio_programme_id.cpp
… src/elements/audio_stream_format.cpp
… src/elements/audio_stream_format_id.cpp
… src/elements/audio_track_format.cpp
… src/elements/audio_track_format_id.cpp
… src/elements/audio_track_uid.cpp
… src/elements/audio_track_uid_id.cpp
… src/elements/format_descriptor.cpp
… src/elements/type_descriptor.cpp
… src/private/copy.cpp
… src/private/xml_parser.cpp
… tests/CMakeLists.txt
… tests/benchmarks.cpp
… tests/helper/parameter_comparators.hpp
… tests/hex_values_tests.cpp