benchmark results

before:
```
common_definitions

benchmark name                       samples       iterations    estimated
                                     mean          low mean      high mean
                                     std dev       low std dev   high std dev
-------------------------------------------------------------------------------
parse                                          100             1    498.248 ms 
                                        4.96482 ms    4.95867 ms    4.97128 ms 
                                        32.3449 us    24.5786 us    45.8501 us 
                                                                               
copy                                           100             1      1.9208 s 
                                         19.247 ms    19.2231 ms    19.2768 ms 
                                        134.811 us    107.649 us    180.961 us 

lots of blocks
-------------------------------------------------------------------------------
/home/thomasn/dev/audio/libadm/tests/benchmarks.cpp:17
...............................................................................

benchmark name                       samples       iterations    estimated
                                     mean          low mean      high mean
                                     std dev       low std dev   high std dev
-------------------------------------------------------------------------------
generate                                        20             1    562.915 ms 
                                        28.1311 ms    28.1025 ms    28.1816 ms 
                                        84.2654 us    47.0731 us    125.765 us 
                                                                               
copy                                            20             1    297.957 ms 
                                        14.9329 ms    14.8969 ms    15.0095 ms 
                                        112.993 us    28.1284 us    175.997 us 
                                                                               
write                                           20             1     6.27171 s 
                                        330.128 ms    327.649 ms    332.466 ms 
                                        5.47824 ms    4.48029 ms    6.79702 ms 
                                                                               
parse                                           20             1     4.05776 s 
                                        191.319 ms    190.105 ms    193.272 ms 
                                        3.46065 ms    2.36129 ms    5.10664 ms 

IDs

benchmark name                       samples       iterations    estimated
                                     mean          low mean      high mean
                                     std dev       low std dev   high std dev
-------------------------------------------------------------------------------
format                                         100            10      1.486 ms 
                                        1.33566 us    1.31723 us    1.42412 us 
                                          177.1 ns    5.93242 ns    422.089 ns 
                                                                               
parse                                          100            21     1.4448 ms 
                                        675.142 ns    669.188 ns    692.235 ns 
                                        46.6849 ns    20.4349 ns    99.6152 ns 

```
after:
```
common_definitions

benchmark name                       samples       iterations    estimated
                                     mean          low mean      high mean
                                     std dev       low std dev   high std dev
-------------------------------------------------------------------------------
parse                                          100             1    223.019 ms 
                                        2.22667 ms    2.22264 ms    2.23109 ms 
                                        21.5264 us       17.2 us     28.132 us 
                                                                               
copy                                           100             1    47.8974 ms 
                                        477.047 us    475.617 us    478.522 us 
                                         7.3977 us    6.48251 us    8.72167 us 

lots of blocks

benchmark name                       samples       iterations    estimated
                                     mean          low mean      high mean
                                     std dev       low std dev   high std dev
-------------------------------------------------------------------------------
generate                                        20             1    574.521 ms 
                                        28.6618 ms     28.615 ms    28.7357 ms 
                                        131.548 us    80.9096 us    218.811 us 
                                                                               
copy                                            20             1    273.487 ms 
                                         13.632 ms    13.6083 ms    13.6963 ms 
                                        83.5453 us    32.1517 us    158.823 us 
                                                                               
write                                           20             1     6.11856 s 
                                        314.262 ms    313.837 ms     314.64 ms 
                                        913.826 us    707.946 us    1.25533 ms 
                                                                               
parse                                           20             1     2.34592 s 
                                        113.089 ms    110.576 ms    116.746 ms 
                                        6.89668 ms    4.61183 ms    8.48877 ms 

IDs

benchmark name                       samples       iterations    estimated
                                     mean          low mean      high mean
                                     std dev       low std dev   high std dev
-------------------------------------------------------------------------------
format                                         100           507     1.4703 ms 
                                        26.7991 ns    26.5985 ns    27.1832 ns 
                                        1.36768 ns   0.855163 ns    2.47528 ns 
                                                                               
parse                                          100           579     1.4475 ms 
                                        26.2201 ns    25.9472 ns    26.5104 ns 
                                        1.43985 ns    1.32877 ns    1.62758 ns 
```

So for documents with lots of references, copying is ~40x faster (mostly from fixing a couple of accidental map copies) and parsing is ~2x faster (mostly from using a map to lookup elements by ID). For documents with a few big elements and no references, there's not much difference, as expected. AudioBlockFormatId parsing and formatting are about 25x and 45x faster too.

There are a lot of changes in here that don't do much (or anything) by themselves, but i think they are worth keeping. Small improvements add up, and having fewer things that look like they might be slow is helpful when looking for real performance issues.

As far as I can see now there's not any obvious performance bottlenecks in the code I've looked at. If parsing needed to go a lot faster I would invert the parsing (so iterate over the elements and decide what to do with them, rather than searching for elements with a given name), because most elements are never used.